Extract FramebufferObject to its own file
