+++ src/celengine/renderglsl.h
@@ -2,7 +2,7 @@
 //
 // Functions for rendering objects using dynamically generated GLSL shaders.
 //
-// Copyright (C) 2006-2009, the Celestia Development Team
+// Copyright (C) 2006-2020, the Celestia Development Team
 // Original version by Chris Laurel <claurel@gmail.com>
 //
 // This program is free software; you can redistribute it and/or
@@ -80,53 +80,4 @@ void renderGeometry_GLSL_Unlit(Geometry* geometry,
                                double tsec,
                                const Renderer* renderer);
 
-
-class FramebufferObject
-{
-public:
-    enum
-    {
-        ColorAttachment = 0x1,
-        DepthAttachment = 0x2
-    };
-    FramebufferObject(GLuint width, GLuint height, unsigned int attachments);
-    ~FramebufferObject();
-
-    bool isValid() const;
-    GLuint width() const
-    {
-        return m_width;
-    }
-
-    GLuint height() const
-    {
-        return m_height;
-    }
-
-    GLuint colorTexture() const;
-    GLuint depthTexture() const;
-
-    bool bind();
-    bool unbind();
-
-
-
-private:
-    void generateColorTexture();
-    void generateDepthTexture();
-    void generateFbo(unsigned int attachments);
-    void cleanup();
-
-private:
-    GLuint m_width;
-    GLuint m_height;
-    GLuint m_colorTexId;
-    GLuint m_depthTexId;
-    GLuint m_fboId;
-    GLenum m_status;
-};
-
-
 #endif // _CELENGINE_RENDERGLSL_H_
-
-
